[#246] Fix No data container handling
We originally used the auth option to allow unowned datasets to be
created in order to allow users to deposit datasets, but that is wrong
because then normal datasets can be created unowed to.

This changes the way auth for depositing datasets is handled and reverts
the auth conf change.