Remove routes to custom endpoints for read dataset / orgs
These were implemented for #131, ie show a Not Authorized error in
dataset and org pages instead of a 404 when a not logged in user tried
to access them. With the latest version of everything this doesn't seem
to be needed anymore, ie trying to access a dataset or org page while
not logged in redirects to the home page with the login links.